+++ src/utils/workspace/workspace_utils.tsx
@@ -1,4 +1,5 @@
-import { App, WorkspaceLeaf, MarkdownView } from "obsidian";
+import { App, WorkspaceLeaf, MarkdownView, editorInfoField } from "obsidian";
+import { EditorView } from "@codemirror/view";
 
 export function getLeafByElement(app: App, el: HTMLElement) : WorkspaceLeaf | null { 
     for (const leaf of app.workspace.getLeavesOfType("markdown")) {
@@ -10,4 +11,11 @@ export function getLeafByElement(app: App, el: HTMLElement) : WorkspaceLeaf | nu
         }
     }
     return null;
+}
+
+export function isSourceMode(view: EditorView): boolean {
+    const mdView = view.state.field(editorInfoField, false) as MarkdownView | undefined;
+    const currentMode = mdView?.currentMode;
+    // @ts-ignore
+    return currentMode?.sourceMode ? true : false;
 }